feat(google_cloud_shelf): extract Shelf serving and logging features into new package
Extract Shelf-specific HTTP serving, contextual logging, trace correlation,
standardized error mapping, and graceful signal termination features from
`pkgs/google_cloud` into a new standalone `pkgs/google_cloud_shelf` package.

- Moved core HTTP serving libraries, middleware, and exceptions:
  - `http_serving.dart` -> `google_cloud_shelf.dart` (package entrypoint)
  - `http_logging.dart` (structured request & trace correlation logging)
  - `http_response_exception.dart` (client-safe structured exceptions)
  - `json_request_checking.dart`
  - `serve.dart` & `terminate.dart` (graceful SIGTERM/SIGINT watch)
- Cleaned up `pkgs/google_cloud`:
  - Removed `shelf` and `stack_trace` dependencies.
  - Deprecated unused `general.dart` exports and removed serving components.
  - Removed unused imports from `test/gcp_test.dart`.
- Migrated and fixed integration tests:
  - Created a new `gcp_test.dart` suite under `pkgs/google_cloud_shelf/test/`
    to maintain full test coverage for port listening, signal termination, and serve handlers.
  - Moved `listen_port_print.dart`, `serve_print.dart`, and `terminate_print.dart`
    helper scripts to the new test location and resolved isolate package URIs.
  - Added `test_process: ^2.0.3` as a `dev_dependency` to `google_cloud_shelf/pubspec.yaml`.
  - Removed the retired serving integration tests from `google_cloud/test/gcp_test.dart`
    to ensure the original package is clean and green.
- Monorepo integration:
  - Added `pkgs/google_cloud_shelf` to the workspace declaration in the root `pubspec.yaml`
    to automatically enable building, testing, and formatting in existing CI workflows.
