+++ pkgs/google_cloud/CHANGELOG.md
@@ -3,41 +3,15 @@ TODO: This is going to be a breaking change release, but we need to bump deps
 -->
 ## 0.4.2-wip
 
+### `google_cloud.dart`: now is the primary exporter of the core APIs.
+
 ### `general.dart`
 
-- REMOVED `CloudLogger`, `LogSeverity`, `structuredLogEntry`.
+- DEPRECATED. Use `google_cloud.dart` instead.
 
 ### `http_serving.dart`
 
-- REMOVED: `CloudLogger`, `LogSeverity`: moved to other packages.
-- REMOVED/BREAKING: `BadRequestException` and `badRequestMiddleware`
-  are removed.
-  - Replaced by `HttpResponseException` and `errorLoggingMiddleware`,
-    respectively.
-- Expanded `HttpResponseException` to support structured error reporting as
-  per AIP-193.
-  - Added `status` (`String?`) and `details` (`List<Map<String, Object?>>?`)
-    fields.
-  - Added `toJson()` method to serialize the error into a standard Google Cloud
-    error payload.
-  - Updated `toString()` to include `status` and `details` when non-null.
-  - Added factory constructors for common HTTP 4XX and 5XX status codes:
-    - `badRequest` (400)
-    - `unauthorized` (401)
-    - `forbidden` (403)
-    - `notFound` (404)
-    - `conflict` (409)
-    - `tooManyRequests` (429)
-    - `internalServerError` (500)
-    - `notImplemented` (501)
-    - `serviceUnavailable` (503)
-    - `gatewayTimeout` (504)
-  - Added default `status` values for factories that map 1:1 to gRPC status
-    codes (e.g., `unauthorized` defaults to `'UNAUTHENTICATED'`).
-- Updated `errorLoggingMiddleware` to leverage
-  `HttpResponseException.toJson()` for JSON responses, returning a standard
-  Google Cloud error payload.
-- Updated plain text errors to use `HttpResponseException.toString()`.
+- REMOVED everything.
 
 ## 0.4.1
