+++ pkgs/google_cloud/pubspec.yaml
@@ -18,8 +18,6 @@ dependencies:
 
 dev_dependencies:
   dart_flutter_team_lints: ^3.2.0
-  googleapis: ^16.0.0
-  googleapis_auth: ^2.0.0
   test: ^1.27.0
   test_descriptor: ^2.0.2
   test_process: ^2.0.3
