+++ pkgs/google_cloud/test/gcp_test.dart
@@ -19,7 +19,6 @@ import 'dart:io';
 import 'dart:isolate';
 
 import 'package:google_cloud/constants.dart';
-import 'package:http/http.dart' as http;
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
 import 'package:test_process/test_process.dart';
@@ -502,143 +501,6 @@ void main() {
       await proc.shouldExit(0);
     });
   });
-
-  group('listenPort', () {
-    late String listenPortPrint;
-
-    setUpAll(() async {
-      final packageUri = await Isolate.resolvePackageUri(
-        Uri.parse('package:google_cloud/'),
-      );
-      listenPortPrint = packageUri!
-          .resolve('../test/src/listen_port_print.dart')
-          .toFilePath();
-    });
-
-    test(
-      'no environment',
-      onPlatform: {
-        'windows': const Skip('TODO: no idea why this is failing on windows'),
-      },
-      () async {
-        final proc = await _run(listenPortPrint);
-
-        await expectLater(proc.stderr, emitsDone);
-        await expectLater(proc.stdout, emits('8080'));
-
-        await proc.shouldExit(0);
-      },
-    );
-
-    test('environment set', () async {
-      final proc = await _run(listenPortPrint, environment: {'PORT': '8123'});
-
-      await expectLater(proc.stderr, emitsDone);
-      await expectLater(proc.stdout, emits('8123'));
-
-      await proc.shouldExit(0);
-    });
-  });
-
-  group(
-    'waitForTerminate',
-    onPlatform: {'windows': const Skip('Cannot validate tests on windows.')},
-    () {
-      late String terminatePrint;
-
-      setUpAll(() async {
-        final packageUri = await Isolate.resolvePackageUri(
-          Uri.parse('package:google_cloud/'),
-        );
-        terminatePrint = packageUri!
-            .resolve('../test/src/terminate_print.dart')
-            .toFilePath();
-      });
-
-      test('sigint', () async {
-        final proc = await _run(terminatePrint);
-
-        await expectLater(proc.stdout, emits('waiting for termination'));
-        await Future<void>.delayed(const Duration(seconds: 1));
-        proc.signal(ProcessSignal.sigint);
-        await expectLater(
-          proc.stdout,
-          emitsInOrder([
-            '',
-            'Received signal SIGINT - closing',
-            'done!',
-            emitsDone,
-          ]),
-        );
-
-        await proc.shouldExit(0);
-      });
-
-      test('sigterm', () async {
-        final proc = await _run(terminatePrint);
-
-        await expectLater(proc.stdout, emits('waiting for termination'));
-
-        await Future<void>.delayed(const Duration(seconds: 1));
-
-        proc.signal(ProcessSignal.sigterm);
-        await expectLater(
-          proc.stdout,
-          emitsInOrder([
-            '',
-            'Received signal SIGTERM - closing',
-            'done!',
-            emitsDone,
-          ]),
-        );
-
-        await proc.shouldExit(0);
-      });
-    },
-  );
-  group(
-    'serveHandler',
-    onPlatform: {'windows': const Skip('Cannot validate tests on windows.')},
-    () {
-      late String servePrint;
-
-      setUpAll(() async {
-        final packageUri = await Isolate.resolvePackageUri(
-          Uri.parse('package:google_cloud/'),
-        );
-        servePrint = packageUri!
-            .resolve('../test/src/serve_print.dart')
-            .toFilePath();
-      });
-
-      test('starts server and responds', () async {
-        final proc = await _run(servePrint, environment: {'PORT': '0'});
-
-        final line = await proc.stdout.next;
-        expect(line, startsWith('Serving at http://'));
-
-        final uri = Uri.parse(line.substring('Serving at '.length));
-
-        final response = await http.get(uri);
-        expect(response.statusCode, 200);
-        expect(response.body, 'Hello from serveHandler!');
-
-        proc.signal(ProcessSignal.sigint);
-
-        await expectLater(
-          proc.stdout,
-          emitsInOrder([
-            '',
-            'Received signal SIGINT - closing',
-            'done!',
-            emitsDone,
-          ]),
-        );
-
-        await proc.shouldExit(0);
-      });
-    },
-  );
 }
 
 Future<TestProcess> _run(
