+++ pkgs/google_cloud/test/metadata_test.dart
@@ -17,7 +17,7 @@ library;
 
 import 'dart:io';
 
-import 'package:google_cloud/general.dart';
+import 'package:google_cloud/google_cloud.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
 import 'package:test/test.dart';
