+++ pkgs/google_cloud/test/src/metadata_host_print.dart
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import 'package:google_cloud/general.dart';
+import 'package:google_cloud/google_cloud.dart';
 
 void main() {
   print(gceMetadataHost);
