+++ pkgs/google_cloud_shelf/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 0.5.0-wip
+
+- Initial release.
