+++ pkgs/google_cloud_shelf/example/contextual_logging.dart
@@ -1,0 +1,45 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// ignore_for_file: cascade_invocations
+
+// #docregion contextual-logging
+import 'package:google_cloud_shelf/google_cloud_shelf.dart';
+import 'package:shelf/shelf.dart';
+
+Future<void> main() async {
+  final handler = const Pipeline()
+      .addMiddleware(createLoggingMiddleware(projectId: 'my-gcp-project-id'))
+      .addHandler(_userHandler);
+
+  await serveHandler(handler);
+}
+
+Response _userHandler(Request request) {
+  // Get the logger specific to this request context
+  final logger = currentLogger;
+
+  logger.info('Fetching user profile from database.');
+
+  // A simple print statement is also captured as an INFO log with trace
+  // correlation
+  print('This print statement is correlated too!');
+
+  // Business logic here...
+  logger.info('User successfully retrieved.', payload: {'userId': 'user_123'});
+
+  return Response.ok('User Profile');
+}
+
+// #enddocregion contextual-logging
