+++ pkgs/google_cloud_shelf/example/error_handling.dart
@@ -1,0 +1,47 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// #docregion error-handling
+import 'package:google_cloud_shelf/google_cloud_shelf.dart';
+import 'package:shelf/shelf.dart';
+
+Response _profileHandler(Request request) {
+  final userId = request.url.queryParameters['id'];
+
+  if (userId == null) {
+    // Automatically returns status 400 to the client
+    throw HttpResponseException.badRequest(
+      message: 'The "id" parameter is required.',
+      details: [
+        {'field': 'id', 'issue': 'must not be empty'},
+      ],
+    );
+  }
+
+  if (userId != 'allowed-user') {
+    // Automatically returns status 403 to the client
+    throw HttpResponseException.forbidden(
+      message: 'Access denied to the requested profile.',
+    );
+  }
+
+  return Response.ok('User Profile Data');
+}
+
+// #enddocregion error-handling
+
+void main() {
+  // Suppress unused element warning for the private handler
+  const _ = _profileHandler;
+}
