+++ pkgs/google_cloud_shelf/lib/google_cloud_shelf.dart
@@ -1,4 +1,4 @@
-// Copyright 2022 Google LLC
+// Copyright 2026 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -22,43 +22,35 @@
 /// ## Example
 ///
 /// ```dart
-/// import 'package:google_cloud/google_cloud.dart';
+/// import 'package:google_cloud_shelf/google_cloud_shelf.dart';
 /// import 'package:shelf/shelf.dart';
 ///
 /// Future<void> main() async {
 ///   final handler = const Pipeline()
 ///       .addMiddleware(createLoggingMiddleware())
-///       .addHandler(_helloWorldHandler);
+///       .addHandler((_) => Response.ok('Hello, World!'));
 ///
 ///   await serveHandler(handler);
 /// }
-///
-/// Response _helloWorldHandler(Request request) {
-///   return Response.ok('Hello, World!');
-/// }
 /// ```
-///
-/// {@canonicalFor http_serving.BadConfigurationException}
-/// {@canonicalFor http_serving.cloudLoggingMiddleware}
-/// {@canonicalFor http_serving.createLoggingMiddleware}
-/// {@canonicalFor http_serving.currentLogger}
-/// {@canonicalFor http_serving.errorLoggingMiddleware}
-/// {@canonicalFor http_serving.HttpResponseException}
-/// {@canonicalFor http_serving.listenPortFromEnvironment}
-/// {@canonicalFor http_serving.serveHandler}
-/// {@canonicalFor http_serving.TraceContextData}
-/// {@canonicalFor http_serving.waitForTerminate}
 library;
 
-export 'src/serving/bad_configuration_exception.dart'
-    show BadConfigurationException;
-export 'src/serving/http_logging.dart'
+export 'src/constants.dart'
+    show
+        cloudTraceContextHeader,
+        defaultListenPort,
+        logSpanIdKey,
+        logTraceKey,
+        logTraceSampledKey,
+        portEnvironmentVariable;
+
+export 'src/http_logging.dart'
     show
         cloudLoggingMiddleware,
         createLoggingMiddleware,
         currentLogger,
         errorLoggingMiddleware;
-export 'src/serving/http_response_exception.dart' show HttpResponseException;
-export 'src/serving/serve.dart' show listenPortFromEnvironment, serveHandler;
-export 'src/serving/terminate.dart' show waitForTerminate;
-export 'src/serving/trace_context_data.dart' show TraceContextData;
+export 'src/http_response_exception.dart' show HttpResponseException;
+export 'src/serve.dart' show listenPortFromEnvironment, serveHandler;
+export 'src/terminate.dart' show waitForTerminate;
+export 'src/trace_context_data.dart' show TraceContextData;
