+++ pkgs/google_cloud_shelf/lib/src/bad_configuration_exception.dart
