+++ pkgs/google_cloud_shelf/lib/src/http_response_exception.dart
