+++ pkgs/google_cloud_shelf/lib/src/json_request_checking.dart
