+++ pkgs/google_cloud_shelf/lib/src/serve.dart
@@ -18,8 +18,8 @@ import 'dart:io';
 import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart';
 
-import '../constants.dart';
 import 'bad_configuration_exception.dart';
+import 'constants.dart';
 import 'terminate.dart';
 
 /// Serves [handler] on [InternetAddress.anyIPv4] using the port returned by
