+++ pkgs/google_cloud_shelf/test/listen_port_print.dart
@@ -1,4 +1,4 @@
-// Copyright 2021 Google LLC
+// Copyright 2022 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,10 +11,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+import 'package:google_cloud_shelf/google_cloud_shelf.dart';
 
-import 'package:google_cloud/google_cloud.dart';
-
-Future<void> main() async {
-  final projectId = await computeProjectId();
-  print('Project ID: $projectId');
+void main() {
+  print(listenPortFromEnvironment());
 }
