+++ pkgs/google_cloud_shelf/test/terminate_print.dart
@@ -11,8 +11,10 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import 'package:google_cloud/google_cloud.dart';
+import 'package:google_cloud_shelf/google_cloud_shelf.dart';
 
-void main() {
-  print(listenPortFromEnvironment());
+Future<void> main() async {
+  print('waiting for termination');
+  await waitForTerminate();
+  print('done!');
 }
