+++ pubspec.yaml
@@ -40,6 +40,7 @@ workspace:
   - generated/google_cloud_type
   - pkgs/google_cloud
   - pkgs/google_cloud_logging
+  - pkgs/google_cloud_shelf
   - pkgs/google_cloud_storage
   - test_utils
   - tests
